requirements.txt
eth-brownie>=1.18.2,<2.0.0
brownie-token-tester>=0.2.2
vyper>=0.3.3
eip712
